+++ design-system/tokens/src/artifacts/figma/atlassian-light/sync-figma-tokens.js
@@ -1426,7 +1426,7 @@ synchronizeFigmaTokens('AtlassianLight', {
       "introduced": "0.6.0",
       "description": "Use as an overlay to communicate selected states when a simple background color change isn't possible, such as in Editor block elements"
     },
-    "value": "#388BFFCC"
+    "value": "#388BFF14"
   },
   "Color/Blanket/Danger": {
     "attributes": {
@@ -1435,7 +1435,7 @@ synchronizeFigmaTokens('AtlassianLight', {
       "introduced": "0.6.0",
       "description": "Use as an overlay to communicate danger states when a simple background color change isn't possible, such as deletion of Editor block elements"
     },
-    "value": "#EF5C48CC"
+    "value": "#EF5C4814"
   },
   "Color/Interaction/Hovered": {
     "attributes": {